requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
curlify==3.0.0
facebook_business==23.0.0
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-bigquery==3.35.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
idna==3.10
multidict==6.6.3
packaging==25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
python-dateutil==2.9.0.post0
requests==2.32.4
rsa==4.9.1
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
